Add tests for percentage
