+++ src/ets/templates/order/detail.html
@@ -35,7 +35,7 @@
       <th class='bordered'>{% trans "Number Of Units" %}</th>
       <th class='bordered'>{% trans "To Be Dispatched" %}</th>
       <th  class='bordered'>{% trans "Stock In Warehouse" %}</th>
-      <th  class='bordered'>{% trans "Percent" %}</th>
+      <th  class='bordered'>{% trans "Percentage" %}</th>
     </tr>
     
     {% for item in object.items.all %}
