+++ src/ets/tests/__init__.py
@@ -5,4 +5,5 @@ from test_waybill import WaybillTestCase
 from test_warehouse import WarehouseTestCase
 from test_api import ApiServerTestCase
 from test_tags import TagsTestCase
+from test_order import OrderTestCase
 #from test_update import UpdateTestCase
