+++ src/ets/tests/test_waybill.py
@@ -62,18 +62,6 @@ class WaybillTestCase(TestCaseMixin, TestCase):
         response = self.client.get(reverse('index'))
         self.assertEqual(response.status_code, 200)
     
-    def test_order_list(self):
-        """ets.views.order_list"""
-        self.client.login(username='dispatcher', password='dispatcher')
-        response = self.client.get(reverse('orders'))
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.context['object_list'].count(), 1)
-
-    def test_order_detail(self):
-        """Order's detail page"""
-        response = self.client.get(reverse('order_detail', kwargs={'object_id': self.order.pk,}))
-        self.assertEqual(response.status_code, 200)
-    
     def test_waybill_view(self):
         """ets.views.waybill_view test"""
         response = self.client.get(reverse('waybill_view', kwargs={'waybill_pk': self.waybill.pk,}))
